refactor: move the directory-scoped routes out of index.ts (#548 step 2a)
Seven routes answer the same question — "tell me about this working directory" —
and each opened with the identical resolve of `?cwd=`: scripts, skills,
dir-config, git-status, pr-phase, header, dir-sound. None of them touch session
state, so they come out first and set the pattern for the rest of step 2.

resolveWorkspace moves to config/workspace.ts alongside a workspaceFromQuery
helper, since every one of these routes re-spelled the same
`typeof req.query.cwd === "string" ? … : null` dance and a route that skipped it
would take an unvalidated path.

Nothing is injected into mountDirRoutes: every dependency was already its own
module once resolveWorkspace moved, so the mount takes only the app.

Behaviour is unchanged, and this time proven rather than reasoned: the five
routes registered after their old neighbours now register earlier, so I booted
main and this branch against the same workspace and diffed the responses of all
seven — byte-identical. index.ts additions are the mount call and its imports.

Refs #548
